# openapi.yaml
openapi: 3.0.3
info:
  title: Bilimcha Media API
  version: "1.0.0"
  description: API для видео, shorts, прокси стриминга, реакций и комментариев
servers:
  - url: http://localhost:5050
paths:
  /api/shorts:
    get:
      summary: Список шортов
      responses:
        '200':
          description: Список шортов
    post:
      summary: Создать short (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
      responses:
        '201':
          description: Short создан
  /api/videos:
    get:
      summary: Список обычных видео
      responses:
        '200':
          description: Список видео
    post:
      summary: Создать видео (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Media'
      responses:
        '201':
          description: Видео создано
...
components:
  schemas:
    Media:
      type: object
      properties:
        id:
          type: integer
        youtubeId:
          type: string
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        thumbnailUrl:
          type: string
        isShort:
          type: boolean
        isPublished:
          type: boolean
        playbackUrl:
          type: string